+++ include/vec/print.h
@@ -1,0 +1,11 @@
+#ifndef VEC_PRINT_H_IS_INCLUDED
+#define VEC_PRINT_H_IS_INCLUDED
+
+#include "vec/types.h"
+
+VecNamespace vec_print (Vec v);
+VecNamespace vec_stats (Vec v);
+
+
+
+#endif