+++ lib/vec.c
@@ -4,6 +4,9 @@
 
 #include "vec/crud.h"
 #include "vec/order.h"
+#include "vec/print.h"
+
+
 
 struct VecNamespace vec_struct = {
 	.new = &vec_new,
@@ -20,6 +23,9 @@ struct VecNamespace vec_struct = {
 
 	.empty = &vec_empty,
 	.full = &vec_full,
+
+	.print = &vec_print,
+	.stats = &vec_stats,
 
 	.swap = &vec_swap,
 	.reverse = &vec_reverse,